+++ .eslintrc.js
@@ -14,7 +14,6 @@ module.exports = {
     ecmaVersion: 12,
     sourceType: "module",
   },
-  plugins: ["jest"],
   rules: {
     "import/no-extraneous-dependencies": [
       "error",
