+++ .github/workflows/github-pages.yml
@@ -12,9 +12,11 @@ jobs:
       - name: Install Packages
         run: |
           npm install
+
       - name: Build
         run: |
           npm run build
+
       - name: Deploy
         uses: JamesIves/github-pages-deploy-action@v4.3.0
         with:
