+++ .gitignore
@@ -5,4 +5,5 @@
 node_modules
 build
 dist
+out
 coverage
