+++ babel.config.js
@@ -1,5 +1,4 @@
 module.exports = (api) => ({
-  plugins: ["@babel/plugin-transform-runtime"],
   presets: [
     [
       "@babel/preset-env",
