+++ lint-staged.config.js
@@ -1,8 +1,5 @@
 module.exports = {
-  "*.js": [
-    "eslint --config eslint.config.js --cache --fix",
-    "prettier --write",
-  ],
+  "*.js": ["eslint --cache --fix", "prettier --write"],
   "*.{css,scss}": "stylelint --cache --fix",
   "*.{json,md,html,css,scss}": "prettier --write",
 };
