+++ package.json
@@ -1,96 +1,86 @@
 {
   "name": "otus-javascript-blog",
-  "version": "1.0.0",
-  "main": "./src/index.js",
   "description": "OTUS JavaScript Blog",
-  "scripts": {
-    "test": "jest",
-    "test:coverage": "jest --coverage",
-    "lint": "eslint . --config eslint.config.js",
-    "lint:fix": "eslint . --config eslint.config.js --fix",
-    "lint:all": "eslint . --config eslint.config.js && stylelint **/*.{css,scss} && prettier . --check",
-    "lint:all:fix": "eslint . --config eslint.config.js --fix && stylelint **/*.{css,scss} --fix && prettier . --write",
-    "prettier": "prettier . --check",
-    "prettier:fix": "prettier . --write",
-    "stylelint": "stylelint **/*.{css,scss}",
-    "stylelint:fix": "stylelint **/*.{css,scss} --fix",
-    "prepare": "husky install",
-    "build": "webpack build --node-env production --config webpack.prod.js",
-    "start": "webpack serve --config webpack.dev.js --mode development",
-    "dev": "webpack serve --config webpack.dev.js --open"
+  "keywords": [
+    "javascript"
+  ],
+  "homepage": "https://github.com/alexey-sidorov-dev/otus-javascript-blog#readme",
+  "bugs": {
+    "url": "https://github.com/alexey-sidorov-dev/otus-javascript-blog/issues"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/alexey-sidorov-dev/otus-javascript-blog.git"
   },
-  "keywords": [
-    "javascript"
-  ],
+  "license": "ISC",
   "author": "Alexey Sidorov",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/alexey-sidorov-dev/otus-javascript-blog/issues"
-  },
-  "homepage": "https://github.com/alexey-sidorov-dev/otus-javascript-blog#readme",
-  "devDependencies": {
-    "@babel/cli": "^7.17.10",
-    "@babel/core": "^7.17.10",
-    "@babel/plugin-transform-runtime": "^7.17.10",
-    "@babel/preset-env": "^7.17.10",
-    "@types/jest": "^27.4.1",
-    "babel-jest": "^28.0.3",
-    "babel-loader": "^8.2.5",
-    "browser-sync-webpack-plugin": "^2.3.0",
-    "copy-webpack-plugin": "^10.2.4",
-    "core-js": "^3.22.3",
-    "css-loader": "^6.7.1",
-    "css-minimizer-webpack-plugin": "^3.4.1",
-    "eslint": "^8.14.0",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jest": "^26.1.5",
-    "fast-glob": "^3.2.11",
-    "handlebars": "^4.7.7",
-    "handlebars-loader": "^1.7.1",
-    "html-loader": "^3.1.0",
-    "html-webpack-plugin": "^5.5.0",
-    "husky": "^7.0.4",
-    "jest": "^28.0.3",
-    "lint-staged": "^12.4.1",
-    "mini-css-extract-plugin": "^2.6.0",
-    "postcss": "^8.4.13",
-    "postcss-import": "^14.1.0",
-    "postcss-loader": "^6.2.1",
-    "postcss-preset-env": "^7.4.4",
-    "prettier": "^2.6.2",
-    "sass": "^1.51.0",
-    "sass-loader": "^12.6.0",
-    "style-loader": "^3.3.1",
-    "stylelint": "^14.8.1",
-    "stylelint-config-prettier": "^9.0.3",
-    "stylelint-config-prettier-scss": "^0.0.1",
-    "stylelint-config-standard": "^25.0.0",
-    "stylelint-config-standard-scss": "^3.0.0",
-    "webpack": "^5.72.0",
-    "webpack-bundle-analyzer": "^4.5.0",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.8.1",
-    "webpack-merge": "^5.8.0"
+  "main": "./src/pages/index.js",
+  "scripts": {
+    "build": "webpack build --define-process-env-node-env production",
+    "dev": "webpack serve --define-process-env-node-env development --open",
+    "lint": "eslint .",
+    "lint:all": "eslint . && stylelint **/*.{css,scss} && prettier . --check",
+    "lint:all:fix": "eslint . --fix && stylelint **/*.{css,scss} --fix && prettier . --write",
+    "lint:fix": "eslint . --fix",
+    "ncu": "npm-check-updates --interactive --format group",
+    "prepare": "husky install",
+    "prettier": "prettier . --check",
+    "prettier:fix": "prettier . --write",
+    "stylelint": "stylelint **/*.{css,scss}",
+    "stylelint:fix": "stylelint **/*.{css,scss} --fix",
+    "test": "jest",
+    "test:coverage": "jest --coverage"
   },
   "dependencies": {
     "modern-css-reset": "^1.4.0"
   },
+  "devDependencies": {
+    "@babel/cli": "^7.21.5",
+    "@babel/core": "^7.21.8",
+    "@babel/preset-env": "^7.21.5",
+    "@types/config": "^3.3.0",
+    "@types/eslint": "^8.37.0",
+    "@types/glob": "^8.1.0",
+    "@types/jest": "^29.5.1",
+    "@types/jsdom": "^21.1.1",
+    "@types/node": "^18.16.3",
+    "babel-jest": "^29.5.0",
+    "babel-loader": "^9.1.2",
+    "browser-sync-webpack-plugin": "^2.3.0",
+    "copy-webpack-plugin": "^11.0.0",
+    "core-js": "^3.30.1",
+    "css-loader": "^6.7.3",
+    "css-minimizer-webpack-plugin": "^5.0.0",
+    "eslint": "^8.39.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jest": "^27.2.1",
+    "handlebars": "^4.7.7",
+    "handlebars-loader": "^1.7.3",
+    "html-loader": "^4.2.0",
+    "html-webpack-plugin": "^5.5.1",
+    "husky": "^8.0.3",
+    "jest": "^29.5.0",
+    "jest-environment-jsdom": "^29.5.0",
+    "lint-staged": "^13.2.2",
+    "mini-css-extract-plugin": "^2.7.5",
+    "npm-check-updates": "^16.10.9",
+    "prettier": "^2.8.8",
+    "prettier-plugin-packagejson": "^2.4.3",
+    "sass": "^1.62.1",
+    "sass-loader": "^13.2.2",
+    "style-loader": "^3.3.2",
+    "stylelint": "^15.6.1",
+    "stylelint-config-standard-scss": "^9.0.0",
+    "webpack": "^5.82.0",
+    "webpack-bundle-analyzer": "^4.8.0",
+    "webpack-cli": "^5.0.2",
+    "webpack-dev-server": "^4.13.3"
+  },
   "overrides": {
-    "browser-sync-webpack-plugin": {
-      "browser-sync": {
-        "portscanner": {
-          "async": "2.6.4"
-        }
-      }
-    },
-    "handlebars-loader": {
-      "async": "2.6.4"
+    "browser-sync": {
+      "ua-parser-js": "1.0.33"
     }
   }
 }
