--- postcss.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  plugins: ["postcss-import", "postcss-preset-env"],
-};
