+++ src/pages/feedback.js
@@ -1,9 +1,9 @@
 import "modern-css-reset";
-import "../styles/scss/main.scss";
+import "../styles/styles.scss";
 
 const page = require("../templates/page.hbs");
 
 const div = document.createElement("div");
 div.className = "content-wrapper";
-div.innerHTML = page({ title: "Список постов" });
+div.innerHTML = page({ title: "Обратная связь" });
 document.querySelector("#app").appendChild(div);
