--- src/pages/index.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <!-- <link
-      rel="shortcut icon"
-      href="../images/favicon.svg"
-      type="image/x-icon"
-    /> -->
-    <title>Personal blog</title>
-  </head>
-  <body>
-    <div id="app"></div>
-  </body>
-</html>
