+++ src/pages/list.js
@@ -1,9 +1,9 @@
 import "modern-css-reset";
-import "../styles/scss/main.scss";
+import "../styles/styles.scss";
 
 const page = require("../templates/page.hbs");
 
 const div = document.createElement("div");
 div.className = "content-wrapper";
-div.innerHTML = page({ title: "Обратная связь" });
+div.innerHTML = page({ title: "Список постов" });
 document.querySelector("#app").appendChild(div);
