+++ src/styles/colors.scss
@@ -1,0 +1,4 @@
+$main-color: #104983;
+$bg-color: rgb(246 248 250);
+$error-color: red;
+$second-color: #404040;
