+++ src/styles/header.scss
@@ -1,0 +1,7 @@
+.menu {
+  &__item {
+    font-family: Arial, Helvetica, sans-serif;
+    text-align: center;
+    text-decoration: none;
+  }
+}
