--- src/styles/scss/fonts.scss
@@ -1,9 +1,0 @@
-@font-face {
-  font-family: "Roboto Regular";
-  src: url("https://fonts.googleapis.com/css2?family=Roboto&display=swap");
-}
-
-@font-face {
-  font-family: "Roboto Bold";
-  src: url("https://fonts.googleapis.com/css2?family=Roboto:wght@700&display=swap");
-}
