--- src/styles/scss/main.scss
@@ -1,38 +1,0 @@
-@import "./fonts";
-
-h1 {
-  font-family: "Roboto Bold", sans-serif;
-  font-size: 36px;
-  text-align: center;
-  margin: 1rem;
-}
-
-header {
-  margin: 1rem;
-}
-
-nav {
-  align-items: center;
-  margin: 1rem;
-  text-align: center;
-  display: flex;
-  flex-direction: column;
-}
-
-.nav__logo {
-  content: url("../../images/logo.svg");
-}
-
-a {
-  font-family: "Roboto Regular", sans-serif;
-  text-align: center;
-  text-decoration: none;
-}
-
-a:visited {
-  color: blue;
-}
-
-a:hover {
-  transform: scale(1.2);
-}
