+++ src/templates/index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="ru">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <meta name="description" content="Персональный блог" />
+    <title>Персональный блог</title>
+  </head>
+  <body>
+    <div id="app" class="app"></div>
+  </body>
+</html>
