+++ src/templates/page.hbs
@@ -1,26 +1,45 @@
-<header>
-  <h1>
-    {{title}}
-  </h1>
-  <nav class="header__nav">
-    <a href="./index.html"><img class="nav__logo" alt="logo" /></a>
-    <a href="./index.html">Персональный блог</a>
-    <a href="./list.html">Список постов</a>
-    <a href="./post.html">Написать пост</a>
-    <a href="./feedback.html">Обратная связь</a>
-    <a href="./about.html">О блоге</a>
-  </nav>
+<header id="header" class="header">
+  <div class="menu">
+    <a href="./index.html"><img class="menu__logo logo" alt="logo" /></a>
+    <nav class="menu__nav nav">
+      <a href="./index.html" class="nav__item">Главная</a>
+      <a href="./list.html" class="nav__item">Посты</a>
+      <a href="./post.html" class="nav__item">Написать пост</a>
+      <a href="./feedback.html" class="nav__item">Обратная связь</a>
+      <a href="./about.html" class="nav__item">О блоге</a>
+    </nav>
+    <div class="menu__social social">
+      <a href="#" aria-label="Telegram" class="social__item"><img
+          class="telegram-icon"
+          alt=""
+        /></a>
+      <a href="#" aria-label="GitHub" class="social__item"><img
+          class="github-icon"
+          alt=""
+        /></a>
+    </div>
+  </div>
 </header>
-<main>
-  <section>
-  </section>
+<main id="main" class="main">
 </main>
-<footer>
-  <nav>
-    <a href="./index.html">Персональный блог</a>
-    <a href="./list.html">Список постов</a>
-    <a href="./post.html">Написать пост</a>
-    <a href="./feedback.html">Обратная связь</a>
-    <a href="./about.html">О блоге</a>
-  </nav>
+<footer id="footer" class="footer">
+  <div class="menu">
+    <nav class="menu__nav nav">
+      <a href="./index.html" class="nav__item">Главная</a>
+      <a href="./list.html" class="nav__item">Посты</a>
+      <a href="./post.html" class="nav__item">Написать пост</a>
+      <a href="./feedback.html" class="nav__item">Обратная связь</a>
+      <a href="./about.html" class="nav__item">О блоге</a>
+    </nav>
+    <div class="menu__social social">
+      <a href="#" aria-label="Telegram" class="social__item"><img
+          class="telegram-icon"
+          alt=""
+        /></a>
+      <a href="#" aria-label="GitHub" class="social__item"><img
+          class="github-icon"
+          alt=""
+        /></a>
+    </div>
+  </div>
 </footer>
